# docs/swagger.yaml
basePath: /
host: tweetgo.vercel.app
info:
  contact: {}
  description: API for scraping X profiles and tweets
  title: X Scraper API
  version: "1.0"
paths:
  /api/v1/profile/{username}:
    get:
      description: Get the profile information of a X user by username
      parameters:
      - description: X username
        in: path
        name: username
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Profile information xScraper.Profile
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid username
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get X profile
      tags:
      - X
  /api/v1/tweets/{username}:
    get:
      description: Get the latest tweets of a X user by username
      parameters:
      - description: X username
        in: path
        name: username
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Latest tweets []xScraper.Tweet
          schema:
            items:
              items:
                additionalProperties:
                  type: string
                type: object
              type: array
            type: array
        "400":
          description: Invalid username
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get X tweets
      tags:
      - X
  /api/v1/xrating/getXIdByUsername:
    get:
      description: Get the ID of an X user by username
      parameters:
      - description: X username
        in: query
        name: username
        required: true
        type: string
      responses:
        "200":
          description: Returns the X ID
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid username
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get X ID by username
      tags:
      - XRating
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to generate token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login to get a JWT token
      tags:
      - Auth
swagger: "2.0"